Move code to copy-mdn.js
