+++ .github/workflows/check.yml
@@ -27,6 +27,9 @@ jobs:
       - name: Install dependencies
         if: steps.caniuse.outputs.newVersion
         run: yarn install --ignore-scripts
+      - name: Copy MDN data
+        id: mdn
+        run: node ./copy-mdn.js
       - name: Pack Can I Use data
         if: steps.caniuse.outputs.newVersion
         run: yarn run pack
